requirements.txt
websockets
numpy
pyopengl
cython
coloredlogs
pyserial
glumpy
glfw